Redesign policy running to be generic with local and server passes
